Finally got notifications to work. A non-swipeable notification shows up during playback and disappears when the app is closed. It has a working play/pause, next, and previous button. Also fixed a few bugs, and did some refactoring.
I used a service to show and hide the notification but it doesn't actually control playback on its own. It communicates with PlayerFragment, which is not a good way to do it but for now it works. We definitely should make a separate class that hosts the mediaPlayer instead of keeping it a static field of a fragment (which right now causes occasional crashes).