Added a SettingsActivity.
For now it's not functional. I wanted to have a setting for library root path  but I couldn't get that to work. I'm not so sure it's what we should do anyway. I have some thoughts I'll put to a new issue. Also I put a setting there for themes but it's not functional yet either.