Fixed issue #77
Created a new method in SongsData which overrides the current queue with a random generated one.  This only gets called when the shuffle checkbox is checked. If it gets unchecked, the queue turns back to the original again.
This highly improves the comfort because you don't get a random generated song every time you go forward or backward in the queue, but only generate it once and can see what comes next.