Fixed bug #98 and playlist refactoring
The bug should now hopefully be fixed. I will test it on my phone too, but it should be ok.
I also had to remake the playlist class as I just found out that the Mediastore playlist class is deprecated and should not be used. They suggest using M3U-files instead so that's what I am going for now.