Made favorite checkbox work
The favorite button now changes look on click and also displays a message. I still need to test the playlist functionality and implement it. I also changed the visualizer color which can now be comfortably changed from the colors.xml file. Furthermore, the SongData class now also has a parameter to lookup if  a song is favorited or not