Playlist is working now
I tested this on an external project and it should be working just fine. Now we just need to implement the Fragment which lists all playlists and we should be good to go. I am just very bad with Fragments so this is going to be fun.
Afterwards we just need to test it and if needed fix any issues.