Fixed issue #131
By adding a custom item value to the themes the background is now also all done and set. Maybe we could add some more colors in the future and probably adjust the color values right now. Also please note that the big round red music circle can't be changed since its a fixed image and it will be replaced by the Song thumbnail in the future.