Improved #132
The previous image has been deleted and has now been replaced by a combined vector asset which can change color according to the selected theme.
I also changed the color of the seekbar from white to the color theme, maybe we should change it back to white or maybe a lighter variant of the theme color.