Preparation for the new release
I think its time we make a new release, the other one is pretty outdated as for now and the other ones are just beta. So as always here is the preparation for the release. It looks like a lot, but they are all actually only minor changes.
I added new backgrounds, a big thanks to @SilviaJack28 for those!
Afterward I also added a changelistener for the settings and changed some colors to fit the new backgrounds.