Moved settings to tabs and removed actionbar for testing prposes
After looking at the new tabs I thought that the actionbar was a little bit too fat. Together with the tabs it used up a lot of the screen and didn't feel that good anymore.
So I commented it out and if you switch to the Red theme you will see the change. I also added the settings to the tab layout since I think having either all in the tabs or all in the dropdown menu just feels more natural and comfortable to me.
This is obviously just my opinion, I would like to hear yours first and see what you think.
Oh and I did try to use the new visualizer and it just looked like we dropped candy in a world where the only color is grey. No but really, it looked completely out of place, didn't have the customization I wanted and therefore had to go immediately. Also, you know, you are right, the visualizer is kind of nice after all.