Actionbr fixed, statusbar not
I am now forcing the actionbar to use the actualy theme color by first retrieving it from the theme, converting it into hex and then setting it. I am sure there is a better way to do this, but I haven't found one. Now I just need to do the same for the statusbar