Adding custom menu
After a lot of thinking I came back with the idea that a custom menu would be quite cool. I already did something like this in the past, but it didn't work out that well. The past version made the main content too small, this version will now be hidden and only shown when needed.