Added new feature
There was a button in the playerfragment, which when clicked only said "Coming soon". Well I changed it and it now opens a new Fragment which displays some information about the song. In the future this will also give users the ability to add tags to songs and order them how they like it best.